refactor: extract re-usable utils for test
